+++ item/model/momo.py
@@ -1,0 +1,23 @@
+from .common import ModelInfo
+
+INFO = ModelInfo(
+    id="momo",
+    citation="""
+    @article{fulton-2009,
+      author = {Lew Fulton and Pierpaolo Cazzola and François Cuenot},
+      doi = {10.1016/j.enpol.2009.07.065},
+      issn = {0301-4215},
+      issuetitle = {Carbon in Motion},
+      journaltitle = {Energy Policy},
+      keywords = {Transport; Modelling; CO2 Emissions},
+      number = {10},
+      pages = {3758–3768},
+      title = {IEA Mobility Model (MoMo) and its use in the ETP 2008},
+      volume = {37},
+      year = {2009},
+      issuesubtitle = {Fuel Economy, Vehicle Use, and Other Factors affecting CO2 Emissions From Transport},
+    }""",
+    format="null",
+    org="International Energy Agency (IEA)",
+    versions=(1, 2),
+)