+++ item/model/roadmap.py
@@ -1,0 +1,18 @@
+from .common import ModelInfo
+
+INFO = ModelInfo(
+    id="roadmap",
+    citation="""
+    @report{facanha-2001
+      author = {Cristiano Façanha and Kate Blumberg and Josh Miller},
+      institution = {International Council on Clean Transportation},
+      location = {San Francisco, CA, US},
+      month = {12},
+      title = {Global transportation energy and climate roadmap},
+      url = {http://www.theicct.org/global-transportation-energy-and-climate-roadmap},
+      year = {2012},
+    }""",
+    format="csv",
+    org="International Council on Clean Transportation (ICCT)",
+    versions=(1, 2, 3),
+)