+++ item/model/shell.py
@@ -1,0 +1,7 @@
+from .common import ModelInfo
+
+INFO = ModelInfo(
+    id="shell",
+    format="null",
+    versions=(2, 3),
+)