+++ item/model/statoil.py
@@ -1,0 +1,7 @@
+from .common import ModelInfo
+
+INFO = ModelInfo(
+    id="statoil",
+    format="null",
+    versions=(2,),
+)